refactor : SOLID-exception
