+++ src/main/java/com/sparta/uglymarket/exception/ErrorResponse.java
@@ -1,0 +1,30 @@
+package com.sparta.uglymarket.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class ErrorResponse {
+    private HttpStatus status;
+    private String message;
+
+    public ErrorResponse(HttpStatus status, String message) {
+        this.status = status;
+        this.message = message;
+    }
+
+    // Getters and Setters
+    public HttpStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(HttpStatus status) {
+        this.status = status;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
