+++ src/main/java/com/sparta/uglymarket/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.sparta.uglymarket.exception;
 
-import com.sparta.uglymarket.exception.CustomException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -26,32 +25,5 @@ public class GlobalExceptionHandler {
     public ResponseEntity<ErrorResponse> handleException(Exception ex) {
         ErrorResponse errorResponse = new ErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, ex.getMessage());
         return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
-
-class ErrorResponse {
-    private HttpStatus status;
-    private String message;
-
-    public ErrorResponse(HttpStatus status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-
-    // Getters and Setters
-    public HttpStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(HttpStatus status) {
-        this.status = status;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
     }
 }
